ajustando regras de erro pra não deslogar adm

+++ plugins/axios.ts
@@ -10,8 +10,12 @@ declare module 'vue/types/vue' {
 
 const api: Plugin = (context, inject) => {
   const api = new Api(context.$axios)
-
   api.setInterceptorResponseError(function (error: any) {
+    if(context.store.state.auth.user.role == 'Administrator') {
+      context.redirect('/queue');
+      alert('possivel erro no url do redis ou redis offline');
+      return Promise.resolve(error);
+    }
     context.store.dispatch('auth/removeSession')
     context.redirect('/')
     return Promise.reject(error);
